# docs/swagger.yaml
openapi: 3.0.2
info:
    title: Games-API
    version: 1.0.0
    description: A brand new API with no content.  Go nuts!
paths:
    /games:
        description: ''
        get:
            tags:
                - games
            responses:
                '200':
                    description: Returns a list of game
            summary: Get a list of all games.
            description: Returns a list of all games in API
        post:
            requestBody:
                description: 'A new game to be created '
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/new_game'
                        examples:
                            Game is created:
                                value:
                                    name: Wow
                                    price: 2.25
                required: true
            tags:
                - new-game
            responses:
                '201':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/new_game'
                    description: Game is created
                '400':
                    description: One or all params are missing
            summary: Creates a game
            description: Creates a new game
    '/games/{id}':
        get:
            tags:
                - details
            responses:
                '200':
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    type: string
                            examples:
                                Full details of a game:
                                    value:
                                        id: 1
                                        name: Witcher 3
                                        price: '29.99'
                    description: Returns full details of a given id game
                '404':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/error'
                            examples:
                                Game /w+ was not found:
                                    value:
                                        error: Game \w+ was not found in game array.
                    description: Game was not found in games array.
            operationId: id
            summary: Get atribfull details of a game
            description: Returns full details of a given id game
components:
    schemas:
        error:
            title: Root Type for error
            description: ''
            type: object
            properties:
                error:
                    type: string
            example:
                error: Error message.
        new_game:
            title: Root Type for new_game
            description: ''
            required:
                - price
                - name
            type: object
            properties:
                name:
                    type: string
                price:
                    format: float
                    type: number
            example:
                name: Tetris
                price: 0.99
    responses:
        200_OK:
            content:
                application/json:
                    schema:
                        type: array
                        items:
                            type: string
                    examples:
                        List of games:
                            value:
                                - Witcher 3
                                - Cyberpun 2077
                                - Wow
            description: Returns a list of games